+++ registry-cli/add-test/digicert_github.crt
