+++ registry-cli/add-test/image-conf-client.json
