+++ registry-cli/add-test/image-conf.json
