+++ registry-cli/add.go
@@ -34,7 +34,7 @@ import (
     "github.com/Multi-Tier-Cloud/common/p2putil"
     "github.com/Multi-Tier-Cloud/common/util"
     driver "github.com/Multi-Tier-Cloud/docker-driver/docker_driver"
-    "github.com/Multi-Tier-Cloud/hash-lookup/hashlookup"
+    "github.com/Multi-Tier-Cloud/hash-lookup/registry"
     "github.com/Multi-Tier-Cloud/service-manager/conf"
 )
 
@@ -164,12 +164,12 @@ func addCmd() {
     }
     defer node.Close()
 
-    info := hashlookup.ServiceInfo{
+    info := registry.ServiceInfo{
         ContentHash: hash,
         DockerHash: dockerId,
         AllocationReq: config.AllocationReq,
     }
-    respStr, err := hashlookup.AddHashWithHostRouting(
+    respStr, err := registry.AddHashWithHostRouting(
         ctx, node.Host, node.RoutingDiscovery, serviceName, info)
     if err != nil {
         log.Fatalln(err)
