+++ registry-cli/add_test.go
