+++ registry-cli/registry-cli.go
