+++ registry-service/handle-add.go
@@ -16,6 +16,7 @@
 
  import (
     "encoding/json"
+    "fmt"
     "io/ioutil"
     "log"
     "strings"
@@ -24,10 +25,10 @@
 
     "go.etcd.io/etcd/clientv3"
 
-    "github.com/Multi-Tier-Cloud/hash-lookup/hl-common"
+    "github.com/Multi-Tier-Cloud/hash-lookup/common"
 )
 
-func handleGet(etcdCli *clientv3.Client) func(network.Stream) {
+func handleAdd(etcdCli *clientv3.Client) func(network.Stream) {
     return func(stream network.Stream) {
         data, err := ioutil.ReadAll(stream)
         if err != nil {
@@ -36,24 +37,25 @@ func handleGet(etcdCli *clientv3.Client) func(network.Stream) {
         }
 
         reqStr := strings.TrimSpace(string(data))
-        log.Println("Lookup request:", reqStr)
+        log.Println("Add request:", reqStr)
 
-        info, ok, err := getServiceInfo(etcdCli, reqStr)
+        var reqInfo common.AddRequest
+        err = json.Unmarshal([]byte(reqStr), &reqInfo)
         if err != nil {
             streamError(stream, err)
             return
         }
 
-        respInfo := common.GetResponse{Info: info, LookupOk: ok}
-        respBytes, err := json.Marshal(respInfo)
+        err = etcdPut(etcdCli, reqInfo.Name, reqInfo.Info)
         if err != nil {
             streamError(stream, err)
             return
         }
 
-        log.Println("Lookup response: ", string(respBytes))
+        respStr := fmt.Sprintf("Added {%s: %v}", reqInfo.Name, reqInfo.Info)
 
-        _, err = stream.Write(respBytes)
+        log.Println("Add response:", respStr)
+        _, err = stream.Write([]byte(respStr))
         if err != nil {
             streamError(stream, err)
             return
