+++ registry-service/handle-delete.go
